requirements.txt
python-telegram-bot[ext]
supabase
openai
python-dotenv
